lib: use u8 for hour
Since the maximum is 99, there is no point using u16.
The SubRip subtitle format only allows for two-digit hours, which
means a subtitle file with an entry that has an 100 hour timecode
will not exist.

+++ src/lib.rs
@@ -7,7 +7,7 @@ pub struct RuntimeArguments {
 }
 
 pub struct TimeCode {
-    pub hour: u16,
+    pub hour: u8,
     pub minute: u8,
     pub second: u8,
     pub millisecond: u16,
@@ -57,7 +57,7 @@ impl RuntimeArguments {
 }
 
 impl TimeCode {
-    pub fn new(hour: u16, minute: u8, second: u8, millisecond: u16) -> TimeCode {
+    pub fn new(hour: u8, minute: u8, second: u8, millisecond: u16) -> TimeCode {
         TimeCode { hour, minute, second, millisecond }
     }
